# solutions/rsw-reference/RSW.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RSW Official",
    "version": "1.0.0",
    "contact": {}
  },
  "paths": {
...
    "/projects/{projectId}/environments/{id}/terminate": {
      "put": {
        "tags": ["environments"],
        "summary": "Terminate Environment",
        "description": "Terminate Environment",
        "operationId": "terminateEnvironment",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of environment to terminate"
        }
      ]
    },
    "/projects/{projectId}/environments": {
      "get": {
        "tags": ["environments"],
        "summary": "List Project Environments",
        "description": "List Project Environments",
        "operationId": "listProjectEnvironments",
        "parameters": [
          {
            "name": "paginationToken",
            "in": "query",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "string",
              "example": ""
            }
          },
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "post": {
        "tags": ["environments"],
        "summary": "Launch Environment",
        "description": "Launch Environment",
        "operationId": "launchEnvironment",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "datasetIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "dataset-123"
                    },
                    "example": []
                  },
                  "description": {
                    "type": "string",
                    "example": "test 123"
                  },
                  "envType": {
                    "type": "string",
                    "example": "sagemakerNotebook"
                  },
                  "envTypeConfigId": {
                    "type": "string",
                    "example": "envTypeConfig-123"
                  },
                  "envTypeId": {
                    "type": "string",
                    "example": "envType-123"
                  },
                  "name": {
                    "type": "string",
                    "example": "testEnv1"
                  }
                }
              },
              "example": {
                "datasetIds": [],
                "description": "test 123",
                "envType": "sagemakerNotebook",
                "envTypeConfigId": "envTypeConfig-123",
                "envTypeId": "envType-123",
                "name": "testEnv1"
              }
            }
          }
        },
        "responses": {
          "201": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        }
      ]
    },
    "/projects/{projectId}/environments/{id}/start": {
      "put": {
        "tags": ["environments"],
        "summary": "Start Environment",
        "description": "Start Environment",
        "operationId": "startEnvironment",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of environment to start"
        }
      ]
    },
    "/projects/{projectId}/environments/{id}/stop": {
      "put": {
        "tags": ["environments"],
        "summary": "Stop Environment",
        "description": "Stop Environment",
        "operationId": "stopEnvironment",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of environment to stop"
        }
      ]
    },
    "/projects/{projectId}/environments/{id}": {
      "get": {
        "tags": ["environments"],
        "summary": "Get Environment",
        "description": "Get Environment",
        "operationId": "getEnvironment",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of environment to GET metadata"
        }
      ]
    },
...
    "/projects/{projectId}/environments/{id}/connections": {
      "get": {
        "tags": ["environments"],
        "summary": "Get Connection",
        "description": "Get Connection",
        "operationId": "getConnection",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          },
          "description": "ID of project the environment belongs to"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID of environment to get connection information"
        }
      ]
    },
...
  },
...
}